Corrections

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: zaddi <zaddi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 11:56:04 by zaddi             #+#    #+#             */
-/*   Updated: 2025/11/13 14:09:29 by zaddi            ###   ########.fr       */
+/*   Updated: 2025/11/14 14:09:57 by zaddi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,62 +27,61 @@ typedef struct s_list
 }					t_list;
 
 // char check function
-int					ft_isascii(int c);
-int					ft_isprint(int c);
-int					ft_isalnum(int c);
-int					ft_isalpha(int c);
-int					ft_isdigit(int c);
-
-int					ft_tolower(int c);
-int					ft_toupper(int c);
+int				ft_isascii(int c);
+int				ft_isprint(int c);
+int				ft_isalnum(int c);
+int				ft_isalpha(int c);
+int				ft_isdigit(int c);
+int				ft_tolower(int c);
+int				ft_toupper(int c);
 
 // Memory function
-void				*ft_calloc(size_t nmemb, size_t size);
-void				*ft_memset(void *s, int c, size_t n);
-void				*ft_memcpy(void *dest, const void *src, size_t n);
-void				*ft_memmove(void *dest, const void *src, size_t n);
-void				ft_bzero(void *s, size_t n);
-void				*ft_memchr(const void *s, int c, size_t n);
-int					ft_memcmp(const void *s1, const void *s2, size_t n);
+void			*ft_calloc(size_t nmemb, size_t size);
+void			*ft_memset(void *s, int c, size_t n);
+void			*ft_memcpy(void *dest, const void *src, size_t n);
+void			*ft_memmove(void *dest, const void *src, size_t n);
+void			ft_bzero(void *s, size_t n);
+void			*ft_memchr(const void *s, int c, size_t n);
+int				ft_memcmp(const void *s1, const void *s2, size_t n);
 
 // String function
-char				*ft_strchr(const char *s, int c);
-char				*ft_strrchr(const char *s, int c);
-char				*ft_strnstr(const char *big, const char *little, int len);
-int					ft_strlcpy(char *dest, const char *src, size_t n);
-int					ft_strlcat(char *dest, const char *src, size_t n);
-int					ft_strlen(const char *str);
-int					ft_strncmp(const char *s1, const char *s2, size_t n);
-char				*ft_strdup(const char *s);
+char			*ft_strchr(const char *s, int c);
+char			*ft_strrchr(const char *s, int c);
+char			*ft_strnstr(const char *big, const char *little, size_t len);
+int				ft_strlcpy(char *dest, const char *src, size_t n);
+int				ft_strlcat(char *dest, const char *src, size_t n);
+int				ft_strlen(const char *str);
+int				ft_strncmp(const char *s1, const char *s2, size_t n);
+char			*ft_strdup(const char *s);
 
-int					ft_atoi(const char *str);
+int				ft_atoi(const char *str);
 
-char				*ft_substr(const char *s, unsigned int start, size_t len);
-void				ft_striteri(char *s, void (*f)(unsigned int, char *));
-char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-char				*ft_strjoin(char const *s1, char const *s2);
-char				*ft_strtrim(char const *s, char const *set);
-char				**ft_split(char const *s, char c);
-char				*ft_itoa(int n);
+char			*ft_substr(const char *s, unsigned int start, size_t len);
+void			ft_striteri(char *s, void (*f)(unsigned int, char *));
+char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+char			*ft_strjoin(char const *s1, char const *s2);
+char			*ft_strtrim(char const *s, char const *set);
+char			**ft_split(char const *s, char c);
+char			*ft_itoa(int n);
 
 // File function
-void				ft_putchar_fd(char c, int fd);
-void				ft_putstr_fd(char const *s, int fd);
-void				ft_putendl_fd(char const *s, int fd);
-void				ft_putnbr_fd(int n, int fd);
-void				ft_putstr_fd(const char *s, int fd);
+void			ft_putchar_fd(char c, int fd);
+void			ft_putstr_fd(char const *s, int fd);
+void			ft_putendl_fd(char const *s, int fd);
+void			ft_putnbr_fd(int n, int fd);
+void			ft_putstr_fd(const char *s, int fd);
 
 // List function
 
-t_list				*ft_lstnew(void const *content);
-void				ft_lstdelone(t_list *alst, void (*del)(void *));
-int					ft_lstsize(t_list *lst);
-void				ft_lstclear(t_list **alst, void (*del)(void *));
-void				ft_lstadd_front(t_list **alst, t_list *new);
-void				ft_lstadd_back(t_list **alst, t_list *new);
-void				ft_lstiter(t_list *lst, void (*f)(void *content));
-t_list				*ft_lstlast(t_list *lst);
-t_list				*ft_lstmap(t_list *lst, void *(*f)(void *),
-						void (*del)(void *));
+t_list			*ft_lstnew(void const *content);
+void			ft_lstdelone(t_list *alst, void (*del)(void *));
+int				ft_lstsize(t_list *lst);
+void			ft_lstclear(t_list **alst, void (*del)(void *));
+void			ft_lstadd_front(t_list **alst, t_list *new);
+void			ft_lstadd_back(t_list **alst, t_list *new);
+void			ft_lstiter(t_list *lst, void (*f)(void *content));
+t_list			*ft_lstlast(t_list *lst);
+t_list			*ft_lstmap(t_list *lst, void *(*f)(void *),
+					void (*del)(void *));
 
 #endif
